Ensure proper handling of edge intervals
Periods of interest that began near the beginning of the week (in UTC
time) were not including intervals from the prior week that overlapped
once time-zone adjustments were taken into account. Vice versa for
backward-looking periods.

This makes it so the starting week is determined by using the time-zone
adjusted origin.