Favor #hours= for interval configuration
This is a small wart that I instantly regretted after putting out
`v1.0.0`. Use of #business_hours= is now considered deprecated and will
be removed in `v2.0.0`. We'll probably put out a migration version
(i.e. `v1.99`) similar to `rspec` that emits warnings for deprecated
methods like this when the time comes to release the next major version.