Make existing `<=>` methods more strict
No more cutesy stuff with duck typing.

After reading a bunch of documentation, including the `Object` docs,
I've come to the conclusion that our use case calls for type checking.
There are some implementations for built-in types that get cute (like
`Fixnum` with coercion), but I don't think any of the `biz` objects
meet the threshold for fanciness like that.

Reference:
* http://ruby-doc.org/core-2.3.0/Object.html#method-i-3C-3D-3E
* https://github.com/rubinius/rubinius/blob/cdc89a9712a910e64b75be4a2e8ed315b487e8ab/core/fixnum.rb#L107-L119