Make value objects full-blown comparable
I swear this is the last time I'm going to tweak this.

The primary reasoning here is that defining `<=>` on `Interval` will
make the sorting that happens during configuration time slightly
cleaner. At that point, there's no reason not to convert over `Holiday`
and `TimeSegment` too, and be done with it once and for all.