Make interval attributes public
This is an interesting one.

It turns out when accessing protected methods on an object of the same
type through the use of `map`, an exception is raised on Ruby 2.2 and
lower.

This seems like a bug and appears to have been fixed in Ruby 2.3.

To avoid this issue, we'll just make the interval attributes public and
document them with tests.