Support zero scalar for-duration calculations
The first active period in the chosen direction is found and the time at
the beginning of that period returned.

If the specified origin is in the middle of an active period, that time
is returned.

The same logic is applied looking forward or backward with the only
difference being the direction.

Fixes #59.