Add holiday performance improvement tip
Calculations will slow down as the number of configured holidays
increases. Unfortunately, there isn't an easy way to optimize for this
in the period generator a priori because we simply don't know when the
generator will stop being asked for periods, and consequently, which
holidays are actually relevant.

Since this could be a concern for some performance-sensitive users,
let's add a tip to the `README` to help them out.

Closes #63.