Filter out empty generated intervals
When short intervals are generated around the DST spring-forward point,
it's possible to generate nonsensical empty intervals because of how the
logic compensates for impossible times during the spring-forward hour by
adding an hour to the time.

Due to this possibility, the period generator should protect against
these empty intervals before emitting a period.