Implement "breaks" (time-segment holidays) feature
Similar to holidays, time during a break is treated as inactive,
regardless of whether there is an active overlapping time segment.

Although this feature is fundamentally similar to holidays, it is
implemented separately in order to keep optimizations related to
holidays that make those calculations approximately 3x faster than
equivalent break-based calculations.

Check out the added specs for a more complete understanding of how this
feature operates in a variety of scenarios.

Closes #64.