Add helper for generating periods on a date
This makes it much more straightforward to answer questions like "What
was the first business time today?" or "What is the last business time
today?".

Having to call `to_a` in order to retrieve the last period on a date is
a bit clunky, but I wanted the API to remain consistent with how
`#before` and `#after` currently work. For the next major version, this
will most likely be changed so all three methods resolve the lazy
enumerator before returning.