Implement shifts
Shifts are date-based exceptions to configured weekly hours.

This has been a frequent feature request (See:
https://github.com/zendesk/biz/issues/5,
https://github.com/zendesk/biz/issues/103,
https://github.com/zendesk/biz/issues/122
).

For a given day in which shifts are configured, intervals that land on that day
will be ignored.

For example, given the following schedule:

```ruby
Biz::Schedule.new do |config|
  config.hours = {
    mon: {'09:00' => '17:00'},
    tue: {'09:00' => '17:00'},
    wed: {'09:00' => '17:00'},
    thu: {'09:00' => '17:00'},
    fri: {'09:00' => '17:00'}
  }

  config.shifts = {
    Date.new(2017, 7, 10) => {'10:00' => '12:00', '13:00' => '16:00'}
  }
end
```

On July 10, 2017, which is a Monday, the base configuration of one active period
from 9am to 5pm will be ignored and two periods of 10am-12pm and 1pm-4pm will be
returned on that date instead.

I decided to retain the optimized version of the `in_hours?` check for
all schedules that do not have shifts configured to maintain performance
parity with competing libraries. In cases where shifts are configured,
the logic drops down to interacting with periods directly in order to
generate a result, which is many multiples slower, but given that
writing an optimized form that included shifts would have turned into
a mess, I think it's a reasonable choice for now.

I added an additional spec to prove the optimized version of `in_hours?`
doesn't work when shifts are configured to guard against future
ill-advised attempts to use that path exclusively.