Enrich the hours validation experience
Previously, the validation logic operated on the raw configuration,
which provides many avenues for invalid configuration to slip through
the cracks unless a litany of cases is tested.

By now having the validation logic operate on the instantiated
configuration object, instead of the underlying raw data, we're able to
catch many more cases of bad configuration with less code that, at the
same time, is more expressive.

Closes #130.