Handle hours like "9:00" with no leading 0
Hours are sometimes written in short format, not being left padded.

In current implementation, this is not recognized:

```
> Biz::DayTime.from_timestamp('01:10:00')
=> #<Biz::DayTime:0x000055a7e6765820 @day_second=4200>

> Biz::DayTime.from_timestamp('1:10:00')
=> nil
```

This forces applications using biz to work around it, and can lead to
production bug if someone working in 24h format only uses afternoon
hours before deploying.

This commit implements support for short format.